requirements.txt
bcrypt==4.0.1
Flask==2.3.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
flask-socketio==5.3.5
flask-cors==4.0.0
flask-login==0.6.2
pymongo==3.12.3
python-dotenv==1.0.0
simple-websocket==0.10.1
python-socketio==5.4.0
python-engineio==4.2.1
Werkzeug==2.3.7
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.2.0
MarkupSafe==3.0.2
flask-pymongo==2.3.0
dnspython==2.1.0
bidict==0.23.1
blinker==1.9.0
